Fix uneven buttons on home page
I'm sure there are lots of clever ways of doing this with CSS (or even
JavaScript), but special thanks to Diego for finding a simpler solution…
Just rewording the stuff in each column. Now, all content takes up the
same number of lines regardless of the breakpoint.

THIS IS MORE OF A HACK THAN A REAL FIX. I'm only doing this because
Design Review is in three days. If the content ever changes, it's just
going to break again.